requirements-test.txt
pytest-django==3.3.2
pytest==4.1.1
